Solve some errors in the pandas stubs
