+++ pandas-stubs/core/frame.pyi
@@ -1,4 +1,3 @@
-import matplotlib
 from pathlib import Path
 from typing import (
     Any,
@@ -14,17 +13,19 @@ from typing import (
     Iterator,
 )
 from typing_extensions import Literal
+import matplotlib
 import numpy as _np
 
-from pandas.core.groupby.generic import DataFrameGroupBy, SeriesGroupBy
-from pandas.core.string import StringMethods
+from .groupby.generic import DataFrameGroupBy, SeriesGroupBy
 from .series import Series
-from .index import Index
+from .indexes import Index
 from .indexing import _iLocIndexer, _LocIndexer
+
+_str = str  # needed because DataFrame has a property called "str"...
 
 _AxisType = Literal["columns", "index"]
 
-_ListLike = Union[_np.ndarray, Series, List, Dict[str, _np.ndarray]]
+_ListLike = Union[_np.ndarray, Series, List, Dict[_str, _np.ndarray]]
 
 _DType = TypeVar("_DType", bound=_np.dtype)
 
@@ -32,7 +33,7 @@ class DataFrame:
     def __init__(
         self,
         data: Optional[Union[_ListLike, DataFrame]] = ...,
-        columns: Optional[Union[List[str], Index]] = ...,
+        columns: Optional[Union[List[_str], Index]] = ...,
         index: Optional[Union[_np.ndarray, Index]] = ...,
     ): ...
     #
@@ -40,18 +41,18 @@ class DataFrame:
     def __eq__(self, other: Union[float, Series, DataFrame]) -> DataFrame: ...
     def __ne__(self, other: Union[float, Series, DataFrame]) -> DataFrame: ...
     @overload
-    def __getitem__(self, idx: str) -> Series: ...
+    def __getitem__(self, idx: _str) -> Series: ...
     @overload
-    def __getitem__(self, idx: Union[List[str], Index]) -> DataFrame: ...
+    def __getitem__(self, idx: Union[List[_str], Index]) -> DataFrame: ...
     def __iter__(self) -> Iterator: ...
     def __len__(self) -> int: ...
     def __setitem__(self, key: Any, value: Any) -> None: ...
     #
     # properties
     @property
-    def columns(self) -> Index[str]: ...
+    def columns(self) -> Index[_str]: ...
     @columns.setter  # setter needs to be right next to getter; otherwise mypy complains
-    def columns(self, cols: Union[List[str], Index[str]]) -> None: ...
+    def columns(self, cols: Union[List[_str], Index[_str]]) -> None: ...
     @property
     def iloc(self) -> _iLocIndexer: ...
     @property
@@ -74,18 +75,18 @@ class DataFrame:
     #
     # methods
     def append(
-        self, s: Union[DataFrame, Dict[str, Any]], ignore_index: bool = ..., sort: bool = ...
+        self, s: Union[DataFrame, Dict[_str, Any]], ignore_index: bool = ..., sort: bool = ...
     ) -> DataFrame: ...
     def apply(self, f: Callable) -> DataFrame: ...
     def copy(self, deep: bool = ...) -> DataFrame: ...
     def count(self) -> Series: ...
-    def drop(self, index: Union[List[str], Index], axis: _AxisType = ...) -> DataFrame: ...
-    def drop_duplicates(self, keep: Union[str, bool] = ...) -> DataFrame: ...
-    def dropna(self, axis: int = ..., how: str = ...) -> DataFrame: ...
+    def drop(self, index: Union[List[_str], Index], axis: _AxisType = ...) -> DataFrame: ...
+    def drop_duplicates(self, keep: Union[_str, bool] = ...) -> DataFrame: ...
+    def dropna(self, axis: int = ..., how: _str = ...) -> DataFrame: ...
     @overload
-    def groupby(self, by: List[str]) -> DataFrameGroupBy: ...
+    def groupby(self, by: List[_str]) -> DataFrameGroupBy: ...
     @overload
-    def groupby(self, by: str) -> SeriesGroupBy: ...
+    def groupby(self, by: _str) -> SeriesGroupBy: ...
     def head(self, n: int) -> DataFrame: ...
     def isnull(self) -> DataFrame: ...
     def max(self) -> Series: ...
@@ -93,13 +94,13 @@ class DataFrame:
     def min(self) -> Series: ...
     def mode(self, axis: _AxisType = ...) -> DataFrame: ...
     def median(
-        self, axis: int = ..., skipna: bool = ..., level: Union[int, str] = ...
+        self, axis: int = ..., skipna: bool = ..., level: Union[int, _str] = ...
     ) -> Union[DataFrame, Series]: ...
     def notna(self) -> DataFrame: ...
     def notnull(self) -> DataFrame: ...
     def nunique(self) -> Series: ...
-    def plot(self, kind: str, yerr: DataFrame) -> matplotlib.axes.Axes: ...
-    def query(self, expr: str) -> DataFrame: ...
+    def plot(self, kind: _str, yerr: DataFrame) -> matplotlib.axes.Axes: ...
+    def query(self, expr: _str) -> DataFrame: ...
     def rename(self, mapper: Callable, axis: _AxisType = ...) -> DataFrame: ...
     def replace(self, a: float, b: float) -> DataFrame: ...
     def reset_index(self, drop: bool) -> DataFrame: ...
@@ -107,22 +108,22 @@ class DataFrame:
     def sample(self, frac: float, random_state: int = ..., replace: bool = ...) -> DataFrame: ...
     @overload
     def sample(self, n: int, random_state: int = ..., replace: bool = ...) -> DataFrame: ...
-    def set_index(self, index: List[str]) -> DataFrame: ...
+    def set_index(self, index: List[_str]) -> DataFrame: ...
     @overload
     def sort_values(
-        self, by: List[str], inplace: Literal[True], axis: _AxisType = ..., ascending: bool = ...
+        self, by: List[_str], inplace: Literal[True], axis: _AxisType = ..., ascending: bool = ...
     ) -> None: ...
     @overload
     def sort_values(
         self,
-        by: List[str],
+        by: List[_str],
         inplace: Optional[Literal[False]] = ...,
         axis: _AxisType = ...,
         ascending: bool = ...,
     ) -> DataFrame: ...
     def std(self) -> DataFrame: ...
     def sum(self) -> Series: ...
-    def to_csv(self, filename: Union[Path, str], index: bool = ...) -> None: ...
+    def to_csv(self, filename: Union[Path, _str], index: bool = ...) -> None: ...
     def to_feather(self, filename: Path) -> None: ...
     @overload
     def to_numpy(self) -> _np.ndarray: ...
