+++ pandas-stubs/core/series.pyi
@@ -2,24 +2,22 @@ from typing import List, Tuple, Type, Union, overload, Optional, Dict, TypeVar, 
 from typing_extensions import Literal
 import numpy as _np
 
-from pandas.core.indexing import _LocIndexer, _iLocIndexer
-from pandas.core.strings import StringMethods
-
-from pandas.core.frame import DataFrame
+from .indexing import _LocIndexer, _iLocIndexer
+from .frame import DataFrame
 from .indexes import Index
 
-class dtype: ...
+_str = str  # needed because Series has a property called "str"...
 
-_DType = TypeVar("_DType", bound=Union[_dtype, int, float, object])
-_DType2 = TypeVar("_DType2", bound=Union[_dtype, int, float, object])
-_ListLike = Union[_np.ndarray[_DType], List[_DType], Dict[str, _np.ndarray[_DType]]]
+_DType = TypeVar("_DType", int, float, object)
+_DType2 = TypeVar("_DType2", int, float, object)
+_ListLike = Union[_np.ndarray, List[_DType], Dict[_str, _np.ndarray]]
 _Num = TypeVar("_Num", int, float, bool)
 
 class Series(Generic[_DType]):
     def __init__(
         self,
         data: Optional[Union[_ListLike[_DType], Series[_DType]]],
-        index: Union[str, int, Series] = ...,
+        index: Union[_str, int, Series] = ...,
     ): ...
     # magic methods
     def __and__(self, other: Series) -> Series: ...
@@ -31,7 +29,7 @@ class Series(Generic[_DType]):
     def __ne__(self, other: object) -> Series: ...
     def __mul__(self, other: float) -> Series: ...
     @overload
-    def __getitem__(self, idx: Union[List[str], Index[int], Series, slice]) -> Series: ...
+    def __getitem__(self, idx: Union[List[_str], Index[int], Series, slice]) -> Series: ...
     @overload
     def __getitem__(self, idx: int) -> float: ...
     def __iter__(self) -> Iterator[_DType]: ...
@@ -70,7 +68,7 @@ class Series(Generic[_DType]):
     def replace(self, to_replace: int, value: int, inplace: bool) -> None: ...
     def std(self) -> float: ...
     def sum(self) -> float: ...
-    def to_dict(self) -> Dict[Union[int, str], _DType]: ...
+    def to_dict(self) -> Dict[Union[int, _str], _DType]: ...
     @overload
     def to_numpy(self) -> _np.ndarray: ...
     @overload
